Fix dedicated tt reweighting in the main executable
Had to split processing of simulation into two distinct steps: SM tt and
other processes. This was needed to avoid reading generator-level
particles for other processes as they are not needed in that case and
for many processes they are not even stored.

Also check in a configuration for application of generator-level
systematic weights in tt.