Always reapply JEC and recompute MET
Recorrect jets and recompute MET not only when individual JEC variations
are evaluated but in any case. This is needed in order to have a
consistent definition of MET. In CMSSW the type-1 correction is
performed with jets that pass a cut on the EM energy fraction and whose
momenta are corrected for the contributing muons. This cannot be
reproduced exactly in mensura. Instead, jets are cleaned against
identified muons and electrons and also rejected if they do not pass the
loose jet ID. This might result in a difference of up to several GeV in
MET, although the same value is reproduced for most of events.

This is an intermediate version that can only be run on simulation.
Different corrections should be applied in data, and their versions
depend on the run number.